Note time in log, change project settings

+++ Amendux/Sha1.cpp
@@ -70,7 +70,7 @@ std::string SHA1::final()
 	}
 
 	/* Append total_bits, split this uint64 into two uint32 */
-	block[BLOCK_INTS - 1] = total_bits;
+	block[BLOCK_INTS - 1] = (int)total_bits;
 	block[BLOCK_INTS - 2] = (total_bits >> 32);
 	transform(block);
